Added a language dropdown.
This is hardcoded to specific languages, but this should be redone to search the rule directory to find the supported languages.

In the JS, Amazon Poly doesn't seem to allow setting the language directly. Instead, you set a voice which speaks a certain language. Again, I had to hard code a mapping from language to voice name which is not good.

Note: there are no voices for Indonesian and Vietnamese. Also, although there is a Finnish voice, it doesn't seem to work.